requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
frozenlist==1.2.0
google-auth==2.3.3
idna==3.3
iso8601==1.0.1
kopf==1.35.3
kubernetes==19.15.0
multidict==5.2.0
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pystache==0.6.0
python-dateutil==2.8.2
python-json-logger==2.0.2
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
typing_extensions==4.0.0
urllib3==1.26.7
websocket-client==1.2.1
yarl==1.7.2
